requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
ansi2html==1.8.0
anyio==3.7.1
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
Babel==2.13.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
click==8.1.7
comm==0.1.4
dataclasses-json==0.6.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
execnb==0.1.5
executing==2.0.0
fastcore==1.5.29
fastjsonschema==2.18.1
fqdn==1.5.1
frozenlist==1.4.0
ghapi==1.0.4
gitdb==4.0.11
GitPython==3.1.40
idna==3.4
ipykernel==6.26.0
ipython==8.16.1
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.2
json5==0.9.14
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.8.0
jupyter-lsp==2.2.0
jupyter_client==8.5.0
jupyter_core==5.4.0
jupyter_server==2.9.1
jupyter_server_terminals==0.4.4
jupyterlab==4.0.7
jupyterlab-pygments==0.2.2
jupyterlab-quarto==0.2.8
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.0
langchain==0.0.324
langchain-experimental==0.0.35
langsmith==0.0.52
loguru==0.7.2
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib-inline==0.1.6
mistune==3.0.2
multidict==6.0.4
mypy-extensions==1.0.0
nbclient==0.8.0
nbconvert==7.9.2
nbdev==2.3.13
nbformat==5.9.2
nest-asyncio==1.5.8
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.26.1
openai==0.28.1
overrides==7.4.0
packaging==23.2
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.11.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.16.1
python-dateutil==2.8.2
python-json-logger==2.0.7
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.4.4
QtPy==2.4.1
referencing==0.30.2
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.6
Send2Trash==1.8.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.22
stack-data==0.6.3
tenacity==8.2.3
terminado==0.17.1
tinycss2==1.2.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.12.0
typer==0.9.0
types-python-dateutil==2.8.19.14
typing-inspect==0.9.0
typing_extensions==4.8.0
unidiff==0.7.5
uri-template==1.3.0
urllib3==2.0.7
watchdog==3.0.0
wcwidth==0.2.8
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
widgetsnbextension==4.0.9
yarl==1.9.2

#-e git+https://github.com/savitharaghunathan/kai.git@e486312d96542c34520a6d507352bf15d7e06b89#egg=kai